% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.data.R
\docType{data}
\name{example.data}
\alias{example.data}
\title{Some example data}
\format{
A list including:
data: The 600 x 20 matrix, the design matrix.
L0: The subgroup to which each sample truly belongs.
Mu0: The true mean parameters of 3 subgroups.
Theta0: The true precision matrices of 3 subgroups.
n_all: The total sample size.
K0: The true number of subgroups 3.
}
\source{
Simulated data (see examples in the function tuning.lambda.FGGM)
}
\description{
Some example data
}
\examples{
data(example.data)
}
