% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genelambda.obo.R
\name{genelambda.obo}
\alias{genelambda.obo}
\title{Generate tuning parameters}
\usage{
genelambda.obo(nlambda1=10,lambda1_max=1,lambda1_min=0.05,
                      nlambda2=10,lambda2_max=1,lambda2_min=0.01,
                      nlambda3=10,lambda3_max=5,lambda3_min=0.5)
}
\arguments{
\item{nlambda1}{The numbers of lambda 1.}

\item{lambda1_max}{The maximum values of lambda 1.}

\item{lambda1_min}{The minimum values of lambda 1.}

\item{nlambda2}{The numbers of lambda 2.}

\item{lambda2_max}{The maximum values of lambda 2.}

\item{lambda2_min}{The minimum values of lambda 2.}

\item{nlambda3}{The numbers of lambda 3.}

\item{lambda3_max}{The maximum values of lambda 3.}

\item{lambda3_min}{The minimum values of lambda 3.}
}
\value{
A sequence of the tuning parameters (lambda1, lambda2, and lambda3).
}
\description{
Generating a sequence of the tuning parameters (lambda1, lambda2, and lambda3).
}
\examples{
lambda <- genelambda.obo(nlambda1=5,lambda1_max=0.5,lambda1_min=0.1, nlambda2=15,lambda2_max=1.5,
                         lambda2_min=0.1, nlambda3=10,lambda3_max=3.5,lambda3_min=0.5)
lambda
}
\author{
Mingyang Ren
}
