% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link_edge_names.R
\name{linked_node_names}
\alias{linked_node_names}
\title{Indexes the names of all nodes connected to some particular nodes in a subgroup.}
\usage{
linked_node_names(summ, va_names, num_subgroup=1)
}
\arguments{
\item{summ}{A list, the summary of the resulting network structures.}

\item{va_names}{A vector, the names of nodes of interest.}

\item{num_subgroup}{Int, the subgroup numbering.}
}
\value{
A list including the names of connected nodes to the nodes of interest in a subgroup.
}
\description{
Indexes the names of all nodes connected to some particular nodes in a subgroup.
}
