% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{plot_network}
\alias{plot_network}
\title{Visualization of network structures.}
\usage{
plot_network(summ, num_subgroup = 1, plot.mfrow,
                    vertex.size=2,vertex.label.cex=0.7,
                    vertex.label.dist=0.75, edge.width = 0.1, l=0)
}
\arguments{
\item{summ}{A list, the summary of the resulting network structures.}

\item{num_subgroup}{Int/vector, the subgroup numbering.}

\item{plot.mfrow}{Figure Layout.}

\item{vertex.size}{The vertex size.}

\item{vertex.label.cex}{The vertex label size.}

\item{vertex.label.dist}{The distance of vertex labels.}

\item{edge.width}{The edge width.}

\item{l}{Node Coordinates.}
}
\value{
Visualization of network structure
}
\description{
Visualization of network structures.
}
