% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_network.R
\name{summary_network}
\alias{summary_network}
\title{The summary of the resulting network structures.}
\usage{
summary_network(opt_Mu_hat, opt_Theta_hat, data)
}
\arguments{
\item{opt_Mu_hat}{A p * K0_hat matrix, the optional mean vectors of K0_hat subgroups.}

\item{opt_Theta_hat}{n * p * K0_hat matrix, the optional precision matrices of K0_hat subgroups.}

\item{data}{A n * p matrix, the design matrix.}
}
\value{
A list including the overlap of edges of different subgroups, the number of edges, and the names of connected nodes to each nodes in each subgroup.
}
\description{
Summarize the characteristics of the resulting network structures.
}
